requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# AI / LLM
openai==1.35.0
langchain==0.2.6
langchain-openai==0.1.14
langgraph==0.1.19

# Vector Database
pinecone-client==4.1.0

# Document Parsing
pymupdf==1.24.5
python-docx==1.1.2

# Data Validation
pydantic==2.7.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
slowapi==0.1.9
tiktoken==0.7.0

# Dev / Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
